test: Adjust tests for recent refactor
